refactor(suite-native): move transaction output labels to transactions
