+++ suite-native/labeling/package.json
@@ -21,13 +21,10 @@
         "@suite-common/suite-sync-storage": "workspace:*",
         "@suite-common/suite-sync-types": "workspace:*",
         "@suite-common/validators": "workspace:*",
-        "@suite-common/wallet-config": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
-        "@suite-common/wallet-types": "workspace:*",
         "@suite-common/wallet-utils": "workspace:*",
         "@suite-native/alerts": "workspace:*",
         "@suite-native/atoms": "workspace:*",
-        "@suite-native/feature-feedback": "workspace:*",
         "@suite-native/forms": "workspace:*",
         "@suite-native/intl": "workspace:*",
         "@suite-native/navigation": "workspace:*",
