+++ suite-native/labeling/src/index.ts
@@ -1,6 +1,4 @@
 export { SendFormLabelEditable } from './components/SendFormLabelEditable';
-export { TransactionOutputLabel } from './components/TransactionOutputLabel';
-export { TransactionOutputLabelEditable } from './components/TransactionOutputLabelEditable';
 export { WalletLabel } from './components/WalletLabel';
 export { EditableLabelLayout } from './components/EditableLabelLayout';
 export { LabelEditForm } from './components/LabelEditForm';
