+++ suite-native/labeling/tsconfig.json
@@ -1,8 +1,12 @@
 {
     "extends": "../../tsconfig.base.json",
-    "compilerOptions": { "outDir": "libDev" },
+    "compilerOptions": {
+        "outDir": "libDev"
+    },
     "references": [
-        { "path": "../../suite-common/device" },
+        {
+            "path": "../../suite-common/device"
+        },
         {
             "path": "../../suite-common/message-system"
         },
@@ -19,27 +23,40 @@
             "path": "../../suite-common/validators"
         },
         {
-            "path": "../../suite-common/wallet-config"
-        },
-        {
             "path": "../../suite-common/wallet-core"
-        },
-        {
-            "path": "../../suite-common/wallet-types"
         },
         {
             "path": "../../suite-common/wallet-utils"
         },
-        { "path": "../alerts" },
-        { "path": "../atoms" },
-        { "path": "../feature-feedback" },
-        { "path": "../forms" },
-        { "path": "../intl" },
-        { "path": "../navigation" },
-        { "path": "../services" },
-        { "path": "../settings" },
-        { "path": "../toasts" },
-        { "path": "../../packages/connect" },
-        { "path": "../../packages/type-utils" }
+        {
+            "path": "../alerts"
+        },
+        {
+            "path": "../atoms"
+        },
+        {
+            "path": "../forms"
+        },
+        {
+            "path": "../intl"
+        },
+        {
+            "path": "../navigation"
+        },
+        {
+            "path": "../services"
+        },
+        {
+            "path": "../settings"
+        },
+        {
+            "path": "../toasts"
+        },
+        {
+            "path": "../../packages/connect"
+        },
+        {
+            "path": "../../packages/type-utils"
+        }
     ]
 }
