+++ suite-native/module-send/package.json
@@ -49,6 +49,7 @@
         "@suite-native/services": "workspace:*",
         "@suite-native/tokens": "workspace:*",
         "@suite-native/transaction-management": "workspace:*",
+        "@suite-native/transactions": "workspace:*",
         "@trezor/blockchain-link-types": "workspace:*",
         "@trezor/connect": "workspace:*",
         "@trezor/react-utils": "workspace:*",
