+++ suite-native/module-send/tsconfig.json
@@ -56,6 +56,7 @@
         { "path": "../services" },
         { "path": "../tokens" },
         { "path": "../transaction-management" },
+        { "path": "../transactions" },
         {
             "path": "../../packages/blockchain-link-types"
         },
