+++ suite-native/transactions/package.json
@@ -19,6 +19,7 @@
         "@suite-common/formatters": "workspace:*",
         "@suite-common/redux-utils": "workspace:*",
         "@suite-common/staking": "workspace:*",
+        "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
         "@suite-common/suite-utils": "workspace:*",
         "@suite-common/token-definitions": "workspace:*",
@@ -29,13 +30,17 @@
         "@suite-common/wallet-utils": "workspace:*",
         "@suite-native/atoms": "workspace:*",
         "@suite-native/device": "workspace:*",
+        "@suite-native/feature-feedback": "workspace:*",
         "@suite-native/formatters": "workspace:*",
         "@suite-native/icons": "workspace:*",
         "@suite-native/intl": "workspace:*",
+        "@suite-native/labeling": "workspace:*",
         "@suite-native/navigation": "workspace:*",
         "@suite-native/scrollview": "workspace:*",
+        "@suite-native/services": "workspace:*",
         "@suite-native/tokens": "workspace:*",
         "@trezor/blockchain-link-types": "workspace:*",
+        "@trezor/connect": "workspace:*",
         "@trezor/styles-native": "workspace:*",
         "@trezor/theme": "workspace:*",
         "@trezor/type-utils": "workspace:*",
