+++ suite-native/transactions/src/components/TransactionOutputLabel.tsx
@@ -2,9 +2,8 @@ import { useSelector } from 'react-redux';
 
 import { type SuiteSyncDataRootState, selectSuiteSyncOutputLabel } from '@suite-common/suite-sync';
 import { Text, type TextProps } from '@suite-native/atoms';
+import { selectIsLabellingAllowed } from '@suite-native/labeling';
 import type { StaticSessionId } from '@trezor/connect';
-
-import { selectIsLabellingAllowed } from '../selectors';
 
 type TransactionOutputLabelProps = TextProps & {
     txId: string;
