+++ suite-native/transactions/tsconfig.json
@@ -8,6 +8,9 @@
         {
             "path": "../../suite-common/redux-utils"
         },
+        {
+            "path": "../../suite-common/suite-sync"
+        },
         { "path": "../../suite-common/staking" },
         {
             "path": "../../suite-common/suite-types"
@@ -35,14 +38,20 @@
         },
         { "path": "../atoms" },
         { "path": "../device" },
+        { "path": "../feature-feedback" },
         { "path": "../formatters" },
         { "path": "../icons" },
         { "path": "../intl" },
+        { "path": "../labeling" },
         { "path": "../navigation" },
         { "path": "../scrollview" },
+        { "path": "../services" },
         { "path": "../tokens" },
         {
             "path": "../../packages/blockchain-link-types"
+        },
+        {
+            "path": "../../packages/connect"
         },
         {
             "path": "../../packages/styles-native"
